Pre Milestone 2

+++ ElariaCogsofTime/FluctLight/Assets/Scripts/SaveData.cs
@@ -4,10 +4,10 @@
 
 public class SaveData
 {
-    public float health;
-    public float currentMove;
+    public int health;
+    public int currentMove;
     public float x;
     public float y;
     public float z;
-    public float currentLevel;
+    public int currentLevel;
 }
